Finally fixed all issues. Will make slides as a report soon. There was actually never a memory issue, when you're not in Debug, memory held from a page is quickly cleared after enteringa new page, as in Debug mode, memory will fill quickly as explained here: https://stackoverflow.com/questions/5299228/django-project-eats-memory
I decided to just use readlines and go backwards because it ensures there is no possible chunking error and that it's actually faster than the chunking method. My analysis was wrong as both were actually O(n^2). Regardless, speed is not an issue as when not in debug mode but have silk enabled, it actually only takes less than 400ms to load the entire full recent activities list. Memory not in debug mode doesn't go over 110 MiB and doesn't stack up.